Add source level coverage feature.
What?
-----
Prior to this feature, only line level coverage existed (mapping of
tests to executable source lines). However, line level coverage tests
run significantly slower than regular runs of the same tests (the
runtime length scales with the number of sources parsed and breakpoints
placed). This runtime performance penalty makes line level coverage
impractical for use cases where large volumes of tests are required to
be run on a regular basis (e.g. in a CI pipeline).

Source level coverage (mapping of tests to source files) provides a
significant performance boost (on par with regular test runs) at the
cost of fidelity. It achieves this by skipping the breakpoint placement
stage of instrumentation and instead assuming that a given enumerated
source file is covered regardless of whether or not any executable
lines of code were covered by the test. This can result in false
positives (sources enumerated as covered but not actually covered) but
will not result in false negatives (sources not enumerated but covered)
any more than line level coverage will.

A new command line option is used to select between source level coverage
and line level coverage: coverage_level

The possible values for this option are source and line (if this option
is not specified the default is line to ensure backwards compatibility
with previous builds).