Add exporter support for source level coverage.
The HTML exporter was having issues exporting coverage data at the
source level due to modules with no line coverage being skipped by the
exporter (source level coverage has no concept of line coverage).

Furthermore, the existing HTML templates had issues rendering the coverage
pie charts due to the paramaters passed in being invalid (both executed
and unexecuted line vars were zero).

These issues have been fixed as follows:
a) A flag is passed in to the HTML exporter to determine whether or not
   modules with no line level coverage will be skipped by the exporter
b) The HMTL templates have had the RGraph instantiation logic replaced
   by a function that instantiated a different pie chart depending on
   whether or not the executed and unexecuted line vars are zero.

The rationale behind this fix was to keep the HTML exporter agnostic
to the concept of line level and source level coverage and thus keep the
changes to this exporter as unintrusive as possible.

The binary and cobertura exporters remain unchanged as they do not have
this issue.